Add warning if a collection is not empty when trying to delete it

+++ Jobsway2goMvc/Controllers/CollectionsController.cs
@@ -204,8 +204,15 @@ namespace Jobsway2goMvc.Controllers
                 return Problem("Entity set 'ApplicationDbContext.Collections'  is null.");
             }
             var collection = await _context.Collections.FindAsync(id);
+            
             if (collection != null)
             {
+                if (collection.Job.Count() > 0)
+                {
+                    ViewBag.DeleteWarning = "The collection you’re about to delete is not empty.";
+                    return View();
+                }
+
                 _context.Collections.Remove(collection);
             }
